refactoring, category and spent integration, removed flyway and introduce sql import in resources

+++ finance-api/src/main/java/com/taurus/financeapi/modules/kitty/repository/KittyRepository.java
@@ -3,5 +3,5 @@ package com.taurus.financeapi.modules.kitty.repository;
 import com.taurus.financeapi.modules.kitty.model.Kitty;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KittyRepository extends JpaRepository<Kitty, Long> {
+public interface KittyRepository extends JpaRepository<Kitty, Integer> {
 }
